backend/ai-agent/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
apscheduler
langchain
langchain-openai
pydantic
python-dotenv
notion-client
langchain-google-genai
Flask
pymongo
cryptography
Flask-Cors
langchain-google-genai
python-dotenv
gtts
openai
